Tuning bytes from ifstream to vector<uint8_t>, add set_offset

+++ io/printer.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 void print_vector(const std::vector<uint8_t>& v) {
     for (auto e : v) {
-        cout << hex << e << " ";
+        cout << hex << static_cast<int>(e) << " ";
     }
     cout << endl;
 }
